python-redis-cluster-sample/requirements.txt
pyyaml==5.3.1
requests==2.25.0
redis-py-cluster==2.1.0